refactor: break large files up (#20)
* fix: split out US and Global APIs

Split out API endpoint functions for US and Global, creating two new files:
 -soilIDModel_US_v3.py
-soilidModel_Global_v3.py

* fix: add awc calc

-Added new functions for simulating AWC

* fix: add info gain function

-added new functions and code handling soil information value and soil variability assessment.

* style: reformat code with black

* remove trailing whitespace
* sort imports

* fix: remove unused imports

* style: remove trailing whitespace

* style: clean up block comment formatting

* style: clean up if True checks

* fix: remove unused exception variables

* fix: generate query string with urllib

* style: wrap long comments

* style: wrap long comments

* style: wrap long SQL queries

* style: wrap long string

* style: wrap long lines

* fix: use params for URL query strings

* style: format with isort/black

---------
